server/requirements.txt
flask>=3.0
flask-cors>=4.0
openpyxl>=3.1
zebra>=0.1.2
python-dotenv>=1.0
